Add files via upload
A mobile classification price project involves building a machine learning model to predict the price category of mobile phones based on various features and attributes. The goal of this project is to create a model that can help users and retailers determine the appropriate price range for a given mobile phone, considering its specifications.